# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CashFlow Manager API
  version: 0.1.0
  description: Production sync API contract for the offline-first CashFlow Manager Flutter app.
servers:
  - url: http://localhost:3000
    description: Local Docker API
security:
  - bearerAuth: []
paths:
...
  /v1/households:
    get:
      summary: List households the current user belongs to
      responses:
        '200':
          description: Household memberships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HouseholdMembershipView'
    post:
      summary: Create a household and owner membership
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HouseholdInput'
      responses:
        '201':
          description: Household created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/households/{id}/invites:
    post:
      summary: Create an owner-only household invite
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HouseholdInviteInput'
      responses:
        '201':
          description: Invite created; token is returned once for local delivery.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdInvite'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /v1/household-invites/{inviteId}/accept:
    post:
      summary: Accept a household invite with its one-time token
      parameters:
        - name: inviteId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptInviteRequest'
      responses:
        '200':
          description: Membership created or restored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HouseholdMember'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/households/{householdId}/members/{memberId}:
    delete:
      summary: Owner-only soft removal of a household member
      parameters:
        - $ref: '#/components/parameters/HouseholdIdPath'
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member removed
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/households/{id}/shared-budgets:
    get:
      summary: List shared budgets in a household
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Shared budget list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SharedBudget'
    post:
      summary: Owner-only create shared budget
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SharedBudgetInput'
      responses:
        '201':
          description: Shared budget created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedBudget'
  /v1/households/{householdId}/shared-budgets/{budgetId}:
    patch:
      summary: Owner-only update shared budget
      parameters:
        - $ref: '#/components/parameters/HouseholdIdPath'
        - name: budgetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SharedBudgetPatchInput'
      responses:
        '200':
          description: Shared budget updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedBudget'
    delete:
      summary: Owner-only soft-delete shared budget
      parameters:
        - $ref: '#/components/parameters/HouseholdIdPath'
        - name: budgetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Shared budget tombstone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedBudget'
...
  /v1/iap/verify:
    post:
      summary: Verify an App Store or Play Store purchase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IapVerifyRequest'
      responses:
        '200':
          description: Purchase verified
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/payments/sepay/orders:
    post:
      summary: Create a direct/off-store SePay order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SePayOrderRequest'
      responses:
        '201':
          description: Payment order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentOrder'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/payments/sepay/webhook:
    post:
      summary: Receive SePay payment status webhook with HMAC signature
      security: []
      parameters:
        - name: x-sepay-signature
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SePayWebhookRequest'
      responses:
        '200':
          description: Webhook accepted
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    HouseholdIdPath:
      name: householdId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Authenticated account lacks permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Optional provider is disabled or unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples: [invalid_amount]
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    SyncMetadata:
      type: object
      required: [id, createdAt, updatedAt, revision]
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: [string, 'null']
          format: date-time
        revision:
          type: integer
          minimum: 1
...
    HouseholdInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    Household:
      allOf:
        - $ref: '#/components/schemas/SyncMetadata'
        - $ref: '#/components/schemas/HouseholdInput'
    HouseholdMember:
      type: object
      required: [id, householdId, userId, role, createdAt]
      properties:
        id:
          type: string
          format: uuid
        householdId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum: [owner, member]
        createdAt:
          type: string
          format: date-time
    HouseholdMembershipView:
      type: object
      required: [role, household]
      properties:
        role:
          type: string
          enum: [owner, member]
        household:
          $ref: '#/components/schemas/Household'
    HouseholdInviteInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        expiresInDays:
          type: integer
          minimum: 1
          maximum: 30
    HouseholdInvite:
      type: object
      properties:
        id:
          type: string
          format: uuid
        householdId:
          type: string
          format: uuid
        inviteeEmail:
          type: string
          format: email
        status:
          type: string
          enum: [pending, accepted, revoked, expired]
        token:
          type: string
          description: Returned only by invite creation.
    AcceptInviteRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
    SharedBudget:
      allOf:
        - $ref: '#/components/schemas/SyncMetadata'
        - $ref: '#/components/schemas/SharedBudgetInput'
    SharedBudgetInput:
      type: object
      required: [name, month, limitAmount]
      properties:
        name:
          type: string
          minLength: 1
        categoryId:
          type: [string, 'null']
          format: uuid
        month:
          type: string
          format: date
        limitAmount:
          type: integer
          minimum: 1
    SharedBudgetPatchInput:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
        categoryId:
          type: [string, 'null']
          format: uuid
        month:
          type: string
          format: date
        limitAmount:
          type: integer
          minimum: 1
...
    IapVerifyRequest:
      type: object
      required: [provider, productId, purchaseToken]
      properties:
        provider:
          type: string
          enum: [apple, google]
        productId:
          type: string
        purchaseToken:
          type: string
        providerSubscriptionId:
          type: string
    SePayOrderRequest:
      type: object
      required: [amount, idempotencyKey]
      properties:
        plan:
          type: string
          default: premium_monthly
        amount:
          type: integer
          minimum: 1
        currency:
          type: string
          enum: [VND]
        idempotencyKey:
          type: string
    PaymentOrder:
      type: object
      required: [id, amount, currency, status, idempotencyKey, createdAt]
      properties:
        id:
          type: string
          format: uuid
        amount:
          type: integer
        currency:
          type: string
        status:
          type: string
          enum: [pending, paid, expired, cancelled]
        providerOrderId:
          type: [string, 'null']
        idempotencyKey:
          type: string
        checkoutUrl:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    SePayWebhookRequest:
      type: object
      required: [providerOrderId, status]
      properties:
        providerOrderId:
          type: string
        status:
          type: string
          enum: [paid, expired, cancelled]
        amount:
          type: integer
        paidAt:
          type: string
          format: date-time